tailwind,head,foot,font

+++ src/main/java/com/koreaIT/demo/controller/UsrArticleController.java
@@ -3,6 +3,7 @@ package com.koreaIT.demo.controller;
 import java.util.List;
 
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
@@ -10,6 +11,8 @@ import com.koreaIT.demo.service.ArticleService;
 import com.koreaIT.demo.util.Util;
 import com.koreaIT.demo.vo.Article;
 import com.koreaIT.demo.vo.ResultData;
+
+import jakarta.servlet.http.HttpSession;
 
 @Controller
 public class UsrArticleController {
@@ -22,7 +25,11 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doWrite")
 	@ResponseBody
-	public ResultData<Article> doWrite(String title, String body) {
+	public ResultData<Article> doWrite(HttpSession session, String title, String body) {
+		
+		if (session.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-L", "로그인 후 이용해주세요");
+		}
 		
 		if (Util.empty(title)) {
 			return ResultData.from("F-1", "제목을 입력해주세요");
@@ -32,24 +39,21 @@ public class UsrArticleController {
 			return ResultData.from("F-2", "내용을 입력해주세요");
 		}
 		
-		articleService.writeArticle(title, body);
+		articleService.writeArticle((int) session.getAttribute("loginedMemberId"), title, body);
 		
 		int id = articleService.getLastInsertId();
 		
 		return ResultData.from("S-1", Util.f("%d번 게시물을 생성했습니다", id), articleService.getArticleById(id));
 	}
 	
-	@RequestMapping("/usr/article/showList")
-	@ResponseBody
-	public ResultData<List<Article>> showList() {
+	@RequestMapping("/usr/article/list")
+	public String showList(Model model) {
 		
 		List<Article> articles = articleService.getArticles();
 		
-		if (articles.size() == 0) {
-			return ResultData.from("F-1", "게시물이 존재하지 않습니다");
-		}
+		model.addAttribute("articles", articles);
 		
-		return ResultData.from("S-1", "게시물 목록", articles);
+		return "usr/article/list";
 	}
 	
 	@RequestMapping("/usr/article/showDetail")
@@ -67,12 +71,16 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(int id, String title, String body) {
+	public ResultData doModify(HttpSession session, int id, String title, String body) {
 		
 		Article article = articleService.getArticleById(id);
 		
 		if (article == null) {
 			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다", id));
+		}
+		
+		if ((int) session.getAttribute("loginedMemberId") != article.getMemberId()) {
+			return ResultData.from("F-A", "해당 게시물에 대한 권한이 없습니다");
 		}
 		
 		articleService.modifyArticle(id, title, body);
@@ -82,12 +90,16 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(int id) {
+	public ResultData doDelete(HttpSession session, int id) {
 		
 		Article article = articleService.getArticleById(id);
 		
 		if (article == null) {
 			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다", id));
+		}
+		
+		if ((int) session.getAttribute("loginedMemberId") != article.getMemberId()) {
+			return ResultData.from("F-A", "해당 게시물에 대한 권한이 없습니다");
 		}
 		
 		articleService.deleteArticle(id);
